renamed files and removed dead code
